Fix bug with Saved & Liked tabs
When I replaced LikedFragment and SavedFragment with common TitlesListTabFragment, it got some problem when it works with "saved" and "liked" collections. The class adds document id as extra for each TitleView. But when it works with "saved" and "liked" collections it should add not document.id but "titleID" field. So I added flag "isSavedTabType" which helps the class to add appropriate ID.